Remove redundant casts

+++ App 112GW/App_112GW/Controls/Buttons/Checkbox.cs
@@ -34,9 +34,9 @@ namespace rMultiplatform
 				HidePoints();
 		}
 		public Checkbox() : base(new SKPoint[3]{
-				new SKPoint((float)(0), (float)(0)),
-				new SKPoint((float)(20), (float)(0)),
-				new SKPoint((float)20, (float)(10))})
+				new SKPoint(0, 0),
+				new SKPoint(20, 0),
+				new SKPoint(20, 10)})
 		{
 			Clicked += CheckboxClick;
 			OffsetAngle = 135;
